+++ Python3/ex12.py
@@ -1,9 +1,9 @@
-#!/bin/python3
+#!/usr/bin/env python3
 
 # ex12: Prompting People
 
 age = input("How old are you?")
-height = input("How tall are you? ")
-weight = input("How much do you weigh? ")
+height = input("How tall are you?")
+weight = input("How much do you weigh?")
 
 print("So, you're %r old, %r tall and %r heavy." % (age, height, weight))
