+++ Python3/ex13-studydrills3.py
@@ -1,4 +1,4 @@
-#!/bin/python3
+#!/usr/bin/env python3
 
 # ex13: Parameters, Unpacking, Variables
 # Combine input with argv to make a script that gets more input from a user.
