+++ Python3/ex13.py
@@ -1,4 +1,4 @@
-#!/bin/python3
+#!/usr/bin/env python3
 
 # ex13: Parameters, Unpacking, Variables
