+++ Python3/ex16-studydrills.py
@@ -13,6 +13,7 @@ print("If you do want that, hit RETURN.")
 input("?")
 
 print("Opening the file...")
+# Open this file in 'write' mode
 target = open(filename, 'w')
 
 print("Truncating the file.  Goodbye!")
@@ -26,12 +27,7 @@ line3 = input("line 3: ")
 
 print("I'm going to write these to the file.")
 
-target.write(line1)
-target.write("\n")
-target.write(line2)
-target.write("\n")
-target.write(line3)
-target.write("\n")
+target.write("%s\n%s\n%s\n" % (line1, line2, line3))
 
 print("And finally, we close it.")
 target.close()
